game-builder-crew/pyproject.toml
[tool.poetry]
name = "game_builder_crew"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
crewai = "^0.85.0"
python-dotenv = "1.0.0"
google-genai = "^1.11.0"
pillow = "^11.2.1"
google-generativeai = "^0.8.5"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[tool.poetry.scripts]
game_builder_crew = "game_builder_crew.main:run"
train = "game_builder_crew.main:train"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
